--- apps/rahat-ui/src/sections/projects/el-kenya/communication/text/text.view.tsx
@@ -1,5 +1,0 @@
-import TextTable from './text.table';
-
-export default function TextView() {
-  return <TextTable />;
-}
