--- apps/rahat-ui/src/sections/projects/el-kenya/communication/text/useTextTableColumn.tsx
@@ -1,37 +1,0 @@
-'use client';
-
-import { Badge } from '@rahat-ui/shadcn/components/badge';
-import { ICampaignItemApiResponse } from '@rahat-ui/types';
-import { ColumnDef } from '@tanstack/react-table';
-
-export default function useTextTableColumn() {
-  const columns: ColumnDef<ICampaignItemApiResponse>[] = [
-    {
-      accessorKey: 'to',
-      header: 'To',
-      cell: ({ row }) => (
-        <div className="cursor-pointer">{row.getValue('to')}</div>
-      ),
-      filterFn: 'includesString',
-    },
-
-    {
-      accessorKey: 'date',
-      header: 'Date',
-      cell: ({ row }) => (
-        <div className="capitalize">{row.getValue('date')}</div>
-      ),
-    },
-    {
-      accessorKey: 'status',
-      header: 'Status',
-      cell: ({ row }) => (
-        <Badge variant="secondary" className="rounded-md capitalize">
-          {row.getValue('status')}
-        </Badge>
-      ),
-    },
-  ];
-
-  return columns;
-}
